@@ -4,9 +4,10 @@
 	"description": "simple react sortable searchable table",
 	"main": "lib/react-sort-search-table.js",
 	"scripts": {
-		"preview": "webpack serve --config=./preview/webpack/webpack.config.js",
-		"compile": "webpack --config=webpack.config.js",
-		"prepublish": "npm run compile"
+		"preview": "webpack serve --open --hot --config=./preview/webpack/webpack.config.js",
+		"compile": "webpack --mode production --config=webpack.config.js",
+		"prepublish": "npm run compile",
+		"test": "jest"
 	},
 	"repository": {
 		"type": "git",
@@ -21,28 +22,60 @@
 		"table"
 	],
 	"peerDependencies": {
-		"react": ">= 16.8.2",
-		"react-dom": ">= 16.8.2"
+		"react": ">= 18.0.0",
+		"react-dom": ">= 18.0.0"
 	},
+	"packageManager": "yarn@3.3.0",
 	"dependencies": {
-		"babel-plugin-styled-components": "^1.12.0",
-		"styled-components": "5.2.3"
+		"@emotion/react": "^11.10.5",
+		"@emotion/styled": "^11.10.5",
+		"@testing-library/dom": "8.20.0",
+		"@testing-library/jest-dom": "5.16.5",
+		"@testing-library/react": "13.4.0",
+		"@testing-library/user-event": "14.4.3",
+		"@types/jest": "29.2.5",
+		"@types/node": "18.11.18",
+		"@types/react": "18.0.26",
+		"@types/react-dom": "18.0.10",
+		"react": "18.2.0",
+		"react-dom": "18.2.0",
+		"typescript": "4.9.3",
+		"web-vitals": "2.1.4"
 	},
 	"devDependencies": {
-		"@babel/core": "7.4.5",
-		"@babel/preset-env": "7.4.5",
-		"@babel/preset-react": "7.0.0",
-		"babel-loader": "8.2.2",
-		"css-loader": "^5.2.0",
-		"html-webpack-plugin": "5.3.1",
-		"react": "16.8.2",
-		"react-dom": "16.8.2",
-		"webpack": "5.30.0",
-		"webpack-cli": "4.6.0",
-		"webpack-dev-server": "3.11.2"
+		"@babel/core": "7.20.5",
+		"@babel/preset-env": "7.20.2",
+		"@babel/preset-react": "7.18.6",
+		"@babel/preset-typescript": "7.18.6",
+		"@emotion/babel-plugin": "^11.10.5",
+		"@types/testing-library__jest-dom": "5.14.5",
+		"babel-loader": "9.1.0",
+		"compression-webpack-plugin": "10.0.0",
+		"css-loader": "6.7.2",
+		"file-loader": "6.2.0",
+		"html-webpack-plugin": "5.5.0",
+		"identity-obj-proxy": "3.0.0",
+		"jest": "29.3.1",
+		"jest-config": "29.3.1",
+		"jest-environment-jsdom": "29.3.1",
+		"pnp-webpack-plugin": "^1.7.0",
+		"style-loader": "3.3.1",
+		"ts-node": "10.9.1",
+		"webpack": "5.75.0",
+		"webpack-cli": "5.0.1",
+		"webpack-dev-server": "4.11.1"
 	},
-	"resolutions": {
-		"styled-components": "^5"
+	"browserslist": {
+		"production": [
+			">0.2%",
+			"not dead",
+			"not op_mini all"
+		],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
 	},
 	"author": "Ching-Ching Yeh <learningg@gmail.com> (https://github.com/Grace951)",
 	"license": "MIT",
