@@ -1,5 +1,5 @@
-import React, { useState, useEffect } from "react";
-import styled from "styled-components";
+import { useState, useEffect } from "react";
+import styled from "@emotion/styled";
 
 const Status = {
 	LOADING: "loading",
