@@ -1,36 +1,18 @@
 const path = require("path");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const projectRoot = process.cwd();
+const PnpWebpackPlugin = require("pnp-webpack-plugin");
 const assetsPath = path.join(projectRoot, "preview", "public");
 const publicPath = "/";
 const host = "0.0.0.0";
 
-const HtmlWebpackPluginConfig = new HtmlWebpackPlugin({
-	title: "ChingChingTest",
-	template: path.resolve(projectRoot, "./preview/index.html"),
-	filename: "index.html",
-	inject: "body",
-});
-
 const config = {
-	devServer: {
-		inline: true,
-		host,
-		port: 3100,
-		hot: true,
-		contentBase: assetsPath,
-		watchOptions: {
-			poll: true,
-		},
-	},
 	cache: false,
 	devtool: "eval",
 	mode: "development",
-	context: projectRoot,
 	entry: {
 		bundle: [path.resolve(projectRoot, "preview", "preview.js")],
 	},
-	target: "web",
 	output: {
 		path: assetsPath,
 		publicPath: publicPath,
@@ -39,31 +21,52 @@ const config = {
 		chunkFilename: "bundle/chunk-[name].js",
 		hotUpdateChunkFilename: "[hash].hot-update.js",
 	},
-	plugins: [HtmlWebpackPluginConfig],
+	devServer: {
+		port: 3100,
+		static: {
+			directory: assetsPath,
+		},
+	},
 	module: {
 		rules: [
 			{
-				test: /(\.jsx)|(\.js)$/i,
-				exclude: [/node_modules/],
-				include: [
-					path.join(projectRoot, "preview"),
-					path.join(projectRoot, "src"),
-				],
+				test: /\.js$|\.tsx?$/,
+				type: "javascript/auto",
+				loader: "babel-loader",
+				exclude: /node_modules/,
+			},
+			{
+				test: /\.css$/i,
 				use: [
+					"style-loader",
 					{
-						loader: "babel-loader",
+						loader: "css-loader",
+						options: {
+							modules: true,
+						},
 					},
 				],
 			},
+			{
+				test: /\.(png|jpg|gif|svg|eot|ttf|woff|woff2|json|xml|ico|cur|ani)$/,
+				use: ["file-loader?name=[path][name].[ext]"],
+			},
 		],
 	},
-	resolveLoader: {
-		modules: [path.join(projectRoot, "node_modules")],
-	},
 	resolve: {
-		modules: [path.join(projectRoot, "node_modules")],
+		plugins: [PnpWebpackPlugin],
 	},
-	profile: true,
+	resolveLoader: {
+		plugins: [PnpWebpackPlugin.moduleLoader(module)],
+	},
+	plugins: [
+		new HtmlWebpackPlugin({
+			title: "ChingChingTest",
+			template: path.resolve(projectRoot, "./preview/index.html"),
+			filename: "index.html",
+			inject: "body",
+		}),
+	],
 };
 
 module.exports = config;
