@@ -1,5 +1,4 @@
-import React from "react";
-import styled from "styled-components";
+import styled from "@emotion/styled";
 
 const Wrap = styled.th`
 	&.defaultCSS {
